Revert "projects: expand_pattern: Do a set difference instead of looping a list"
This partially reverts commit a6d7d53ee716664169a66d060322707e1d269371.

The reason for this revert is that glob pattern expansion code in
util.py works with sets to ensure uniqueness of the partial matches
collection. Sets are unordered which poses a problem for the 'build'
command where the order of projects specified on the command line must
be honored.
This is only one part of the fix, the other will have to make sure glob
pattern matching code is only run when there's actually something to
expand, i.e. don't run it on verbatim lists.

 Conflicts:
    guests/lcitool/lcitool/projects.py

    - missing context: try-except block was added around the util
                       version of the expand_pattern method
    - variable rename: the original variable holding the result was
                       named 'projects' which was later changed to
                       'projects_expanded' which better reflects the
                       reality

Signed-off-by: Erik Skultety <eskultet@redhat.com>