lcitool: Drop wrappers around Inventory.facts
First of all, the Inventory.has_host() function is no longer
accurately named because it checks whether a target, not a
host, is known to lcitool.

Moreover, since we're exposing the full dictionary already, it
doesn't make a lot of sense to have these very thin wrappers
around it and it's overall better to simply let the calling code
use the standard Python dictionary APIs, even if it result in a
small amount of code duplication.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>