lcitool: Introduce Inventory.host_facts
Instead of extracting just a plain list of hosts from the
user-provided inventory (as parsed by Ansible), we can go a bit
further and also collect the corresponding host facts.

This enables a couple of future enhancements (inventory
validation and better user experience for 'lcitool install'),
as well as being more ergonomic because it closely mirrors what
already happens with Inventory.target_facts.

Inventory.hosts, like Inventory.targets, is now just a convenience
property that returns the list of keys in the corresponding fact
dictionary.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>