guests: application: Check for missing optional dependencies early
If one doesn't need to create nor manage VMs, requiring both Ansible
and ansible-runner is quite a stretch. Currently we check that the
ansible-runner package is installed whenever the AnsibleWrapper class
is instantiated. That's too late and we should know long before we
enter that code whether we have all we need to proceed.
The way this patch implements the early dependency checking is by
introducing a decorator that tries to import all dependencies a
subcommand defines it will need.

Signed-off-by: Erik Skultety <eskultet@redhat.com>