lcitool: Treat AnsibleWrapper module more like a plugin
If the user doesn't wish to work with VMs, we should not require the
Ansible bits to be install. Consequently, we should not import anything
that requires the Ansible bits on the global level. Therefore, move the
AnsibleWrapper class imports to places which make use of the class.
This effectively replaces commit a6f873e6.

Signed-off-by: Erik Skultety <eskultet@redhat.com>