guests: drop git commit information from file headers
When the warning header about files being auto-generated was added, the
git commit hash of libvirt-ci.git was included. The idea was that it
shows that version of the package mappings / project package list was
used. This is well intentioned but in practice has caused a number of
problems, mostly around synchronization. A typical sequence is

 - Modify libvirt-ci.git to add a new package
 - Run lcitool in $PROJECT.git to re-generate dockerfiles
 - Submit a merge request for libvirt-ci
 - Submit a merge request for $PROJECT

The problem here is that when the merge request for libvirt-ci.git
is approved, it often ends up being rebased before being applied.
Each time it is rebased, the dockerfiles need to be re-generated
otherwise they're pointing to a commit that no longer exists in the
public repo. This rebasing need is both tedious and easily forgotten.

The git commit hash info is not a net-win in the end. If projects
really do want to keep track of the libvirt-ci.git hash used for
their genreated files, it is probably better if they attach
libvirt-ci.git as a git submodule instead.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>