lcitool: Make util.expand_pattern() preserve ordering
Most of the time, the order in which matching items are returned
doesn't matter, but in the case of the "build" action we actually
need to build projects in a certain order or things will simply
not work.

Let's always return matches in a way that preserves order:
callers that don't care won't be worse off, and those that do
will appreciate the change.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>