lcitool: Simplify wildcard rejection for build command
The string "all" as an alias for "*" is only supported when it's
the entirety of the pattern to be expanded, not as a sub-pattern,
and we want to avoid "*" being present altogether. As a result,
we don't need to check sub-patterns separately and we can just
look at the pattern in its entirety to decide whether or not we
should allow it.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>