lcitool: Start separating internal projects
Right now, both projects that are intended to be visible to the
user and internal projects are treated the same: this makes some
amount of sense, but also results in unfortunate things such as
the need to filter the latter out in certain scenarios.

Start separating the two by moving YAML files for internal
projects to a subdirectory. Later patches will further improve
on the separation.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>