util: Introduce a cross platform validation helper
Currently we check whether a specific OS target supports cross
compilation in the formatters module right before dumping a Dockerfile.
We'll want to perform the same kind of check for the 'packages' command
as well. Therefore, extract the check to a standalone helper and
perform the validation at a common place for both callers, i.e. when
trying to resolve the project generic package names.

Signed-off-by: Erik Skultety <eskultet@redhat.com>