tests: Add unit tests for the package resolving code
There are essentially only 4 test cases which are then parametrized
with the supported targets to ensure the resolver is tested on all of
the platforms.
The generic package name resolution is performed over all of the
mappings we currently recognize in lcitool and is reflected by the
input dataset 'packages_in.yml' which all of the test cases use.

Signed-off-by: Erik Skultety <eskultet@redhat.com>