tests: Drop the 'data_dir' fixture
We need to know the path to the test data dir in order to regenerate
any expected test outputs. The problem is that fixtures cannot be
called directly, so let's convert the fixture to a constant.
This simple solution will do fine for a while, but eventually we'll
have to create a test package so that we can have test util data and
make imports still work.

Signed-off-by: Erik Skultety <eskultet@redhat.com>