tests: Drop the 'facts' fixture
Since we implemented inventory as a singleton we don't need fixtures to
cache any data for us, it is cached by our implementation of singleton,
although it's a tiny bit slower (metaclass dictionary lookup). In any
case it helps to decrease the number of unnecessary fixtures in this
test module.

Signed-off-by: Erik Skultety <eskultet@redhat.com>