tests: Introduce a helper which converts a list of raw packages to dict
Currently we use sets to compare the actual and expected data
iterables, because it's convenient. The problem is that the current
code doesn't play nicely with the idea of supporting output
regeneration which is in YAML where we need a dictionary.
This patch converts the initial list of raw packages to a dictionary
using the package type (native, cross, etc.) as keys and stores the
corresponding package names as sorted lists. Comparing sorted lists is
equally good as sets (if not better) and opens us the door towards
introducing output regeneration code.

Signed-off-by: Erik Skultety <eskultet@redhat.com>