guests: fix dockerfile layering for ccache / package list commands
The package list query is supposed to record all packages installed in
the container. This needs to be run after the cross-architecture
packages have been installed when relevant.

The ccache symlinks contain architecture specific components in their
name, so must be associated with the cross-architecture commands when
relevant. This currently breaks caching of the first command block
layer since it is not arch agnostic.

For both these cases we have a set of commands that either need to be
in the first command block (native case) or the second command block
(cross arch case).

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>