tests: Improve dictionary comparison
The two dictionaries are supposed to be identical, and if they
differ in any way we want an error to be raised.

The comparison in test_package_resolution() is not strict enough,
because it silently ignores the case where expected contains more
data than actual.

The one in test_verify_all_mappings_and_packages() doesn't suffer
from this issue, but comparing the two dictionaries as a whole
results in errors that look like

  AssertionError: assert {'native': ['...-devel', ...]} == {'native': ['...-devel', ...]}
    Differing items:
    {'native': ['SDL2-devel', 'SDL2_image-devel', 'alsa-lib-devel', 'ant', 'ant-junit',
     'audit-libs-devel', ...]} != {'native': ['SDL2-devel', 'SDL2_image-devel',
     'alsa-lib-devel', 'ant', 'ant-junit', 'audit-libs-devel', ...]}

whereas comparing keys one at a time yields the much more
friendly

  AssertionError: assert ['SDL2-devel'...s-devel', ...] == ['SDL2-devel'...s-devel', ...]
    At index 14 diff: 'bzip2-devel' != 'ca-certificates'
    Left contains one more item: 'zlib-static'

so we use the latter.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>