mappings: Improve mapping for netcf
netcf is in the process of being dropped from many (most?)
Linux distributions.

We've already dealt with its disappearance from Debian sid in
746697f4fd69 but the implementation was not future-proof: in
particular, we'd have to touch the mapping again once Debian 12
is out. And since Ubuntu is a Debian derivative, it's fair to
assume that their next LTS release is not going to have netcf
either.

On the RPM front, CentOS Stream 9 will not include netcf and
OpenSUSE never did in the first place.

The plan for Fedora is also to drop the package, but since
both Fedora 35 and Fedora Rawhide still include it updating
that part of the mapping feels premature. We have no idea
what Alpine is going to do about netcf, so that part is left
alone for now as well.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>