guests: add support for local data dir overrides
This adds support for a new arg to all commands taking project names

  -d DIR / --data-dir DIR

as a way to provide local overrides for project package definitions.

Given such an arg, it will attempt to load projects matching the
pattern:

   $DIR/projects/*.yml

Any projects loaded which have the same filename as a built-in project
will take precedence.

Note the arg is registered globally, rather than against each command,
so that it can be consistently applied to all, even if some don't
currently make use of it.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>