formatters: Change distro update strategy for rolling RPM distros
Rolling distros sometimes remove packages (deliberately, by mistake,
for testing) or downgrade them. Simple package upgrade won't work in
this case because the package manager cannot cope with resolving broken
dependencies. Therefore, the package managers often provide a more
intelligent way of upgrading by syncing all packages to versions
currently installed repositories provide, including removing packages
for which an upgrade would not be possible. We're already using
dist-upgrade on Debian and Ubuntu, this is just a partial extension of
the idea to RPM distros.
For example OpenSUSE documents that distro sync type of upgrade is the
only recommended one for Tumbleweed.

Resolves: https://gitlab.com/libvirt/libvirt-ci/-/issues/8

Signed-off-by: Erik Skultety <eskultet@redhat.com>