guests: ansible: Log the dumped group vars file contents
Because all of the Ansible magic happens in a temporary directory,
we don't have any trace of the actual data that was passed to Ansible,
because the directory hierarchy will have been cleaned up by then.
Note that this deliberately doesn't log extra vars, because those
can contain sensitive info.

Signed-off-by: Erik Skultety <eskultet@redhat.com>