guests: support multilayer cross dockerfiles
The caching of native commands in cross dockerfiles is not especially
effective, because the native commands are cached separately for each
cross dockerfile.

We need to support creating a standalone base dockerfile containing
only the native pieces and ten inherit from that for the cross
dockerfile. This guarantees we only process the native pieces
once, no matter how many cross dockerfiles we create.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>