mappings: Simplify some mappings
In a few cases, we can significantly reduce the complexity of
mappings at the cost of introducing a little duplication. This
is a good compromise, because it makes it less likely that bugs
will be introduced in the mappings file due to not realizing
that, for example, multiple overrides with increasing
specificity exist.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>