lcitool: Add CentOS Stream 9
CentOS Stream 9 is mostly the same as CentOS Stream 8, with a
few notable differences:

  * the PowerTools repository is now called CRB;

  * the Advanced Virtualization repository no longer exists;

  * several -devel packages, as well as a few Python and misc
    packages, have been dropped.

In addition to the above, there are also a couple of
differences that are only specific to the early life of CentOS
Stream 9 and should go away in due time:

  * EPEL is not yet available;

  * a small handful of packages that have been built for
    CentOS Stream 9 are still not included in the repositories
    for whatever reason.

All the changes needed because of the above are clearly marked
with special comments in the mappings file, so that once CentOS
Stream 9 is more mature we can easily go back and drop them.

Note that in a lot of cases, when a package that was present in
CentOS 8 and CentOS Stream 8 is no longer present in CentOS
Stream 9, the corresponding mapping is modified to look like

  foo:
    rpm: foo-devel
    CentOS:
    CentOS8: foo-devel
    CentOSStream8: foo-devel

instead of the more obvious

  foo:
    rpm: foo-devel
    CentOSStream9:

This is done intentionally to make the mapping future proof:
since the package was actively removed from the distro, it's
fair to assume that it will not just come back in the next
release, and so having older releases rather than the newest
one, which we expect to set the trend for future ones, as the
exception makes more sense.

These exceptions will naturally be dropped as we drop support
for the corresponding release too, so the mapping will become
simpler over time.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>