guests: drop lttng-ust from FreeBSD platform
The package is no longer installable on either FreeBSD 12 or 13,
with the error message:

  pkg: No packages available to install matching 'lttng-ust'
       have been found in the repositories

Investigation shows this is a result of a failed build for lttng-ust
followed by a push of the repo contents which dropped it, rather than
simply using previous build. It has been broken for a few weeks with
no sign of being fixed. Since lttng-ust is at best a 'nice to have'
for FreeBSD, dropping it is a prudent option to get CI working again.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>