guests: prevent test_commands.py leaving behind files
The 'test_commands.py' test invokes the lcitool commands and has some
logic to cleanup files created by the 'manifest' command. This cleanup,
however, is broken as the location of the output is relative to the
current directory from which the command is launched, not relative to
the location of the test code.

Switching to run with the '-n' command gets the manifest tool to stop
creating stuff on disk and operate in "dry run" mode instead, avoiding
the need for any cleanup.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>