guests: allow manifest command output directories to be changed
Currently we unconditionally output in the "ci" directory underneath the
current working directory.

Some projects may wish to use a directory different that the top level
'ci' directory,  so allow this to be customized via '--ci-dir PATH'.
This should be a path relative to the projet root directory, as it is
also used in the GitLab CI config file to access resources.

When invoking the command it may also be desirable to explicitly set
the base directory rather than rely on the current working directory,
so that behaviour is predictable regardless of execution context.
Allow this to be controlled via '--base-dir PATH'.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>