guests: Install the EPEL repo on CentOS Stream 9
The repo has been added since we introduced the distro to lcitool.
Contrary to previous CentOS releases, EPEL is currently no longer
installed through the 'epel-release' package, instead, it is installed
directly as an RPM from [1] according to the official documentation
[2].

Note, that some of the packages in our mappings we expect to be part of
EPEL have still not been added to EPEL 9.

[1] https://dl.fedoraproject.org/pub/epel/
[2] https://docs.fedoraproject.org/en-US/epel/#_el9

Signed-off-by: Erik Skultety <eskultet@redhat.com>