guests: formatters: Extract Dockerfile string alignment to a method
We never hit an issue with the alignment, because we only ever needed
to align package lists for the given project, but never for any of the
commands that prepare the container instance to do something with the
project - until EPEL 9 has entered the equation.

A more robust and generic solution to handle proper string alignment
would be perfect, but a simple standalone helper method will do fine
for now.

Signed-off-by: Erik Skultety <eskultet@redhat.com>