gitlab: switch to trigger jobs against the merge request by default
Currently we run a pipeline on every push event. These run pre-merge
in the context of the contributor's fork, or post-merge in the context
of the upstream project.

GitLab introduced a rather low limit for CI minutes for new contributors
which makes it essentially impossible for them to run the CI pipelines
for libvirt projects. Since we require a passing pipeline before
accepting merge requests, this is a contribution barrier.

If we run CI jobs against the merge request, then in cases where the
contributor can't run their own pipeline, a project member can
trigger a pipeline on their behalf in the context of upstream.

We don't want to run all pipelines in upstream context as that will
consume too many CI minutes from the project.

Users might sometimes wish to run a pipeline on push as in the past.
We allow this to be manually requested by setting $LIBVIRT_CI=1 for
the pipeline. e.g.

   git push -o ci.variable=LIBVIRT_CI=1

The final complication is that we still need to publish containers
post-merge, so there is a rule allowing build and push of the containers
when pushing to the default branch only. In the context of the merge
request we only want to build the container, not push it to the
registry, to avoid polluting the registry with tags for merge requests.

A side effect of this is that in the merge request, we need to directly
run against the check-dco content we have, not the pre-built container
which might be older. Not using the pre-built container doesn't cost
us much since the only extra package we need to install is git.

In summary:

 - If a member of the project opens a merge request a pipeline runs
   in the context of the upstream immediately
 - If a non-member of the project opens a merge request, a pipeline
   attempts runs in the context of their fork. If they don't have
   ability to run pipelines, a project member must manually trigger
   a pipeline from the "Pipelines" UI
 - Containers are built on push to 'master'
 - If pushing to a branch with LIBVIRT_CI=1 a pipeline is run

The env variable naming is chosen with a view to future policy. The
intent is that we have a few groupings

 * CI_xx - standard GitLab CI vars exposed to jobs
 * CIRRUS_xx - vars for influencing Cirrus CI integration
 * LIBVIRT_CI_xx - vars for influencing libvirt pipelines, controlled
                   by contributors via git push options or similar
 * LIBVIRT_JOB_xx - vars for influencing execution rules in templates,
                    defined by libvirt jobs

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>