gitlab: skip container jobs if not applicable changes are present
Now that we are only running container build jobs on push post-merge
or in merge requests, we can reliably match on the commit file paths
to control when the job is invoked.

Thus we set rules to ensure that containers are only built if there
is a change to the gitlab config file or the container source files.

This will avoid wasting CI time building containers when we're only
changing the lcitool source.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>