guests: playbooks: update: Import EPEL's GPG key explicitly
DNF in its default configuration will import EPEL's GPG key when the
repo RPM is installed. That however is not the case with Ansible's DNF
module which also sets the 'localpkg_gpgcheck' option which would
result in a failure to install any package for which we're missing the
corresponding GPG key.
This patch makes sure we import EPEL 9's GPG key before installing the
repo from RPM.

Signed-off-by: Erik Skultety <eskultet@redhat.com>