mappings: Add perl ExtUtils-CBuilder mapping
We've been relying on the package to be installed as a weak dependency
of perl Module-Build. On CentOS Stream 9 container the dependency
resolver fails to include ExtUtils-CBuilder as a dependency to
Module-Build without a prior DNF update. In any case to solve this
problem it's better to introduce an explicit mapping.

Signed-off-by: Erik Skultety <eskultet@redhat.com>