gitlab-ci: Mark centos-stream-9 as allow-failure
Recent builds of the CentOS Stream 9 container image caused a CI
failures across all our libvirt project umbrella and there's no ETA
when the CentOS team is going to address the issue.
In the meantime, mark the distro as allow-failure so as not to remain
blocked on unrelated codebase changes.

Signed-off-by: Erik Skultety <eskultet@redhat.com>