guests: commandline: Add '--target' install option
In order to allow quick and convenient host installs without the need
to specify the host names in an inventory file, we need to know what
target OS is the <host> going to run, otherwise we wouldn't know which
facts to use to correctly manage it. This patch adds a commandline
option for that.

Signed-off-by: Erik Skultety <eskultet@redhat.com>