guests: inventory: Cache the unified Ansible inventory
Another preparation patch before introducing support for convenient
installation of local hosts which would have libvirt XML metadata
marked by lcitool.
Once a host is installed and marked and the user runs 'lcitool update'
on it, we first need to gather host facts for which we execute Ansible
for the first time to fetch inventory data. Then, when everything is
ready, we execute Ansible for the second time, this time with a
playbook.
The thing is that a clean runner environment was used for
each Ansible execution and while the former did count in the hosts
defined in libvirt locally, the latter didn't because it lacked context.
To fix that we'd have to execute Ansible for a third time to give us
the inventory again so that it would include the locally installed
hosts right before executing the playbook. A better approach is to
simply cache the Ansible inventory result (which doesn't change within
a single lcitool session) and re-use that.

Signed-off-by: Erik Skultety <eskultet@redhat.com>