guests: application: Support '--target' with the install command
Previously we only supported hosts explicitly recorded in an inventory
file. For locally installed hosts, we only needed to check that each of
the hosts was defined with the 'fully_managed=True' inventory variable.
Now that lcitool generates an additional inventory for new locally
installed libvirt hosts, we also need to make sure that the '--target'
install option doesn't collide with the existing 'fully_managed=True'
check.

Signed-off-by: Erik Skultety <eskultet@redhat.com>