guests: lcitool: Replace all generic Exceptions with ValueError
While it's okay to catch *all* exceptions in a generic manner, it's
hardly ever good practice to raise generic exceptions, because the
class inheritance could result in many exceptions being chained together
before presented to the user. Additionally, by using different/custom
exception types we have the flexibility of handling individual
exceptions.
In context of this patch, all occurrences of raising the generic
Exception type fits the ValueError exception type perfectly. At the same
time none of the occurrences is or should be a user-facing error and
hence should be wrapped by a module-level exception which is directly
presented to the user.

Signed-off-by: Erik Skultety <eskultet@redhat.com>