Drop the python3-selinux package from the vm project package listing
It turns out, having this as part of the package installation phase is
too late as we use the 'template' module in the 'base.yml' tasks
list which is one of the early task list executed on a host.
Next patch will take care of installing python3-selinux early on
SELinux-enabled platforms.

This partially reverts commit 5fce605b.

Signed-off-by: Erik Skultety <eskultet@redhat.com>