guests: ansible_wrapper: Don't duplicate stdout on playbook error
The main problem is that ansible-runner <2.0.0 pipes both stderr and
stdout to a single stream, so we don't have a way to dump the error
stream in case of an execution error. The thing is that we run
ansible-runner with different workloads, so sometimes we silence the
output and sometimes we let it through to the console. If a verbose
execution failed, then given the current way of handling errors, stdout
would be dumped along with an error message which effectively would
duplicate the already existing console output.

Therefore, dump the stdout for error reporting purposes only for
Ansible commands we run silenced (i.e. using params['quiet']). In all
other cases report a simple error that an Ansible command failed
assuming that information about the failure had been dumped to stdout
by Ansible itself.

Suggested-by: Andrea Bolognani <abologna@redhat.com>
Signed-off-by: Erik Skultety <eskultet@redhat.com>