lcitool/manifest: split gitlab container template jobs into separate file
Currently ci/gitlab.yml contains all the generated build logic.

It is desired to be able to only rebuild containers when either their
dockerfile changes, or the gitlab build logic changes. While we could
write a rule matching ci/gitlab.yml that will trigger redundant
container rebuilds when simply changing the set of targets, or
build jobs.

This change splits the container template jobs into a separate file
called 'ci/gitlab/container-templates.yml'

With this split it is possible to declare:

   rules:
     - changes:
         ci/gitlab/container-templates.yml
	 ci/containers/$NAME.Dockerfile

to optimize exactly when a specific container is rebuilt.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>