guests: add 'xz' for libvirt-php, libvirt-sandbox & libvirt-cim projects
The automake 'distcheck' target accidentally passes the '--best' arg to
gzip when decompressing, when it only makes sense for compressing. This
is harmless with the normal gzip impl, but triggers an error with the
busybox gzip.

We have been relying on getting 'gzip' indirectly via an existing pkg
dependency, and recently this changed from the full gzip to a busybox
minimal gzip in openSUSE tumbleweed.

While we could add an explicit gzip dep, we instead take the opportunity
to switch remaining projects over to 'xz'.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>