lcitool: Drop cargo-clippy gitlab job
Turns out that this is not really usable in its current form.

First of all, running clippy requires building the project, and
so having a separate job for it running in a fresh container
means that you're wasting a lot of CPU cycles.

More importantly, projects like libvirt-rust need a number of
extra packages such as the native library, pkg-config, libclang
and so on to be available in addition to the Rust compiler and
clippy, which makes providing a one-size-fits-all container
pretty much impossible.

Reverts: 22b83c9516047fd2fa6f640d928efc3348799916
Signed-off-by: Andrea Bolognani <abologna@redhat.com>