lcitool/manifest: delay combining of cidir and basedir paths
The _generate_gitlab method is going to need access to paths
relative to basedir, so that it can generate the include
statements for the gitlab config. The includes must contain
cidir, but not contain basedir.

In order to achieve this, we don't want to combine cidir and
basedir in the constructor, but instead delay it until time
comes to write out the files / create the dirs.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>