lcitool: Tweak `bin/lcitool` to ensure execution works from git root
This is to prepare for the removal of the ./lcitool shell wrapper script.

The ./lcitool script allows execution of lcitool without prior package
installation by `pip install -e .` or `python setup.py develop`.
It does this by adding lcitool/ to PYTHONPATH.

This patch provides the same functionality by modifying `bin/lcitool`
which will in turn allow us to drop the wrapper script.

Signed-off-by: Abdulwasiu Apalowo <abdulwasiuapalowo@gmail.com>