guests: fix allow-failure handling for containers
A container for a given (target, arch) combination can be
used by multiple build jobs. When emitting the container
job definition we just used the properties from the first
build job. Thus if the first build job is set to allow
failure but the second build job is mandatory, we incorrectly
set the container job to allow failure.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>