guests: Application: Use the now mandatory --osinfo virt-install option
Since virt-install-4.0.0 (Fedora 36) the '--osinfo' option is now
required and if missing the execution fails with:

    ERROR
    --os-variant/--osinfo OS name is required, but no value was
    set or detected.

    This is now a fatal error. Specifying an OS name is required
    for modern, performant, and secure virtual machine defaults.

Since lcitool VM's are rather single purpose we don't really miss out
on much by not selecting a specific OS variant with virt-install.
After all, this project and osinfodb introduce OS support at different
pace and we'd be locking ourselves by waiting for the host distro to
release an updated version of osinfodb (if ever on older host OSes)
just to get support for a platform we want to test on.
Therefore, the easiest fix for the option is to pass the 'unknown'
OS target to virt-install.

Signed-off-by: Erik Skultety <eskultet@redhat.com>