formatters: Fix use of projects in build_varmap()
The data returned by Project.get_packages() is not generated on
the fly, but rather cached internally by the Project instance to
avoid having to recalculate it multiple times per run.

As a consequence, we have to make sure that we treat it as
read-only, otherwise subsequent calls to the function will yield
different results.

We can see the behavior in action in the test cases that are
updated in this commit: calling

  pytest.param("test-ccache", "debian-10", "s390x",
               id="ccache-debian-10-cross-s390x")

causes the list of packages for the "base" internal project to
be updated to include ccache, and so the call to

  pytest.param("test-minimal", "debian-10", "s390x",
               "debian-10-common", "all",
               id="minimal-debian-10-common-cross-s390x")

which happens later outputs that package even though it should
not.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>