guests: mark disabled jobs as manual rather than omitting
Currently if a job has 'builds: false', we entirely omit the
definition of the build job, but keep the container job.

This is inconvenient if a developer wants to manually test
a job. With this change, we now always create all container
and build jobs.

If 'builds: false', then the build job is marked as manual.

The manual jobs are visible in the pipeline but don't
contribute to the result, and are never run unless the
developers requests it explicitly.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>