lcitool: block pipelines from triggering against merge requests
While we would like to have pipelines associated with merge requests,
the way we publish container builds for use in later jobs in the
pipeline is not compatible with this. It would end up publishing
conmtainers to the upstream repo for merge requests that we not yet
approved.

As such we need to explicitly block all jobs from running in a merge
request scenario, for now.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>