inventory: Don't error out when the inventory file or dir is missing
The new style of installation will make sure that we'll always have at
least an empty dictionary inventory (from _get_libvirt_inventory)
passed onto Ansible and so it will never complain about anything.
Before this new style of handling a "dynamic" inventory we had to
explicitly check that an inventory source was present to return a nice
error to the user. Given the above this is no longer needed and it
actually improves user experience when running lcitool for the first
time.

Signed-off-by: Erik Skultety <eskultet@redhat.com>