commandline: Edit the install command's host's help
The current wording "host to act on" while accurate is quite vague in
how it describes both scenarios where a host entry may already exist in
the inventory file and we just want to install it again OR we're
installing a completely new host without prior addition to the
inventory file (libvirt's metadata will take care of that).
Reword the help slightly to hint that either an existing entry can be
re-used or a new name should be selected. While at it, rename the
hostopt to installhostopt as only the install command is acting upon a
single host, all other commands either take a list or act on targets.

Signed-off-by: Erik Skultety <eskultet@redhat.com>