mappings: Replace bare FreeBSD keyword occurrences with 'pkg'
Since the pkg format is represented only by FreeBSD, we can use the
package format qualifier just like we do with rpm or deb, etc. instead
of specifying the OS distro name where the mapping applies to all
supported versions. We've been quite inconsistent and so this patch
takes care of that.

Signed-off-by: Erik Skultety <eskultet@redhat.com>