lcitool: README: Delete this README
Historically, we've had multiple READMEs in the file which confused
some users when looking for proper documentation. With all information
extracted from this file and split into standalone RST documents which
can later be made into a proper web hosted documentation, there's no
need for this file any longer and can be dropped.

Signed-off-by: Erik Skultety <eskultet@redhat.com>