mappings: add mapping for glib-mkenums
HomeBrew split all the python code out of the main 'glib'
package, so any project that wants python devel tools like
'glib-mkenums' on macOS needs to install 'glib-utils' too.

Debian previously also split it into a -bin package,
though that is a dep from the -dev package, so we've not
needed to specify it explicitly before. An explicit
mapping, however, means we are guaranteed to get the
native version in cross builds.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>