lcitool: move multiple key lookup to Package superclass
The logic to look up multiple mappings is common to both native
and cross packages, and we will also extend it to PyPI and CPAN
in order to disable some packages.

Move it up to the superclass so that it can be reused; because
subclasses do not return MappingKeyError, the exception goes
away altogether.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>