lcitool: allow disabling pypi and cpan packages
Right now, it is not possibly to disable pypi and cpan packages.
However, this means that whenever a Python (or Perl, though that is
really done only for completeness) package is not available as a native
package, lcitool needs to know about all -devel packages for all of its
dependencies; otherwise, pypi fails to build the package's non-Python
parts.

This may be very complicated to do if the distro only has a very small
number of Python packages available to install; this is the case for
RHEL's Python modules (e.g. python38) and for OpenSUSE's auxiliary
Python runtimes (e.g. python39).  The simplest thing is to just punt
and block those packages; lcitool will not be able to build dockerfiles
that require them, but otherwise everything will work.

To achieve this effect, look up all the keys even for pypi and cpan
packages.  An empty key will cause the package not to be installed.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>