lcitool: add more pypi packages
flake8 is not part of CentOS 9 and AlmaLinux 9 but it can be installed
from pypi.

docutils is also a Python package, so add it to the mappings too.
For now it is just for completeness, but in the future it will be
useful for CentOS 8's modular installation of Python 3.8, or for
OpenSUSE Leap 15.3's Python 3.9 packages.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>